Break tests apart
